requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.7
langchain-community==0.2.7
langchain-huggingface==0.0.3
chromadb==0.5.3
sentence-transformers==2.7.0
google-generativeai==0.8.2
openai==1.35.7
pypdf==4.3.1
PyPDF2==3.0.1
python-multipart==0.0.9
reportlab==3.6.13
python-docx==1.1.2
fpdf==1.7.2
